# libs/domains/agenstra/backend/feature-agent-controller/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Controller API
  version: 1.0.0
  description: HTTP API for managing clients and proxied agent operations
  contact:
    name: ForePath
    email: hi@forepath.io
    url: https://forepath.io
servers:
  - url: http://localhost:3100/api
    description: Local development
  - url: http://localhost:3100
    description: Root (OpenTelemetry metrics; outside /api prefix)
security:
  - bearerAuth: []
paths:
...
  /filter-rules:
    get:
      summary: List regex filter rules (admin only)
      operationId: listFilterRules
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
          description: Maximum number of rules to return
        - in: query
          name: offset
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of rules to skip (pagination)
        - in: query
          name: search
          schema:
            type: string
          description: Case-insensitive search across pattern, filterType, direction, replaceContent, id
      responses:
        '200':
          description: Rules with per-workspace sync summary
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FilterRuleResponseDto'
        '403':
          description: Forbidden
    post:
      summary: Create filter rule (admin only)
      operationId: createFilterRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFilterRuleDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilterRuleResponseDto'
        '400':
          description: Validation error
        '403':
          description: Forbidden
  /filter-rules/{id}:
    get:
      summary: Get filter rule (admin only)
      operationId: getFilterRule
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Rule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilterRuleResponseDto'
        '403':
          description: Forbidden
        '404':
          description: Not found
    put:
      summary: Update filter rule (admin only)
      operationId: updateFilterRule
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFilterRuleDto'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilterRuleResponseDto'
        '400':
          description: Validation error
        '403':
          description: Forbidden
        '404':
          description: Not found
    delete:
      summary: Delete filter rule (admin only)
      operationId: deleteFilterRule
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '403':
          description: Forbidden
        '404':
          description: Not found
  /imports/atlassian/connections:
    get:
      summary: List Atlassian site connections (admin only)
      operationId: listAtlassianSiteConnections
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
          description: Maximum number of connections to return
        - in: query
          name: offset
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of connections to skip (pagination)
        - in: query
          name: search
          schema:
            type: string
          description: Case-insensitive search across label, baseUrl, accountEmail, id
      responses:
        '200':
          description: Connections (API tokens never returned)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AtlassianSiteConnectionResponseDto'
        '403':
          description: Forbidden
    post:
      summary: Create Atlassian site connection (admin only)
      operationId: createAtlassianSiteConnection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAtlassianSiteConnectionDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlassianSiteConnectionResponseDto'
        '400':
          description: Validation error
        '403':
          description: Forbidden
  /imports/atlassian/connections/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get Atlassian site connection (admin only)
      operationId: getAtlassianSiteConnection
      responses:
        '200':
          description: Connection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlassianSiteConnectionResponseDto'
        '403':
          description: Forbidden
        '404':
          description: Not found
    put:
      summary: Update Atlassian site connection (admin only)
      operationId: updateAtlassianSiteConnection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAtlassianSiteConnectionDto'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlassianSiteConnectionResponseDto'
        '403':
          description: Forbidden
        '404':
          description: Not found
    delete:
      summary: Delete Atlassian site connection (admin only)
      operationId: deleteAtlassianSiteConnection
      responses:
        '204':
          description: Deleted
        '403':
          description: Forbidden
        '404':
          description: Not found
  /imports/atlassian/connections/{id}/test:
    post:
      summary: Test Atlassian connection (admin only)
      operationId: testAtlassianSiteConnection
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlassianConnectionTestResultDto'
        '403':
          description: Forbidden
  /imports/atlassian/configs:
    get:
      summary: List external import configs (admin only)
      operationId: listExternalImportConfigs
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
          description: Maximum number of configs to return
        - in: query
          name: offset
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of configs to skip (pagination)
        - in: query
          name: search
          schema:
            type: string
          description: Case-insensitive search across provider, importKind, jql, cql, confluenceSpaceKey, id
      responses:
        '200':
          description: Configs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExternalImportConfigResponseDto'
        '403':
          description: Forbidden
    post:
      summary: Create import config (admin only)
      operationId: createExternalImportConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExternalImportConfigDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalImportConfigResponseDto'
        '400':
          description: Validation error
        '403':
          description: Forbidden
  /imports/atlassian/configs/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get import config (admin only)
      operationId: getExternalImportConfig
      responses:
        '200':
          description: Config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalImportConfigResponseDto'
        '403':
          description: Forbidden
        '404':
          description: Not found
    put:
      summary: Update import config (admin only)
      operationId: updateExternalImportConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExternalImportConfigDto'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalImportConfigResponseDto'
        '403':
          description: Forbidden
        '404':
          description: Not found
    delete:
      summary: Delete import config (admin only)
      operationId: deleteExternalImportConfig
      responses:
        '204':
          description: Deleted
        '403':
          description: Forbidden
        '404':
          description: Not found
  /imports/atlassian/configs/{id}/run:
    post:
      summary: Run import once for config (admin only)
      operationId: runExternalImportConfig
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '202':
          description: Accepted
        '403':
          description: Forbidden
  /imports/atlassian/configs/{id}/markers:
    delete:
      summary: Clear sync markers for config (admin only)
      operationId: clearExternalImportMarkers
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Cleared
        '403':
          description: Forbidden
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
...
  schemas:
...
    TicketStatus:
      type: string
      enum: [draft, todo, in_progress, prototype, done, closed]
...
    FilterRuleSyncSummaryDto:
      type: object
      required: [pending, synced, failed]
      properties:
        pending:
          type: integer
          minimum: 0
        synced:
          type: integer
          minimum: 0
        failed:
          type: integer
          minimum: 0
    FilterRuleWorkspaceSyncDto:
      type: object
      required: [clientId, syncStatus]
      properties:
        clientId:
          type: string
          format: uuid
        syncStatus:
          type: string
          enum: [pending, synced, failed]
        lastError:
          type: string
    FilterRuleResponseDto:
      type: object
      required:
        [
          id,
          pattern,
          regexFlags,
          direction,
          filterType,
          priority,
          enabled,
          isGlobal,
          workspaceIds,
          sync,
          workspaceSync,
          createdAt,
          updatedAt,
        ]
      properties:
        id:
          type: string
          format: uuid
        pattern:
          type: string
        regexFlags:
          type: string
        direction:
          type: string
          enum: [incoming, outgoing, bidirectional]
        filterType:
          type: string
          enum: [none, filter, drop]
        replaceContent:
          type: string
        priority:
          type: integer
        enabled:
          type: boolean
        isGlobal:
          type: boolean
        workspaceIds:
          type: array
          items:
            type: string
            format: uuid
          description: Empty when isGlobal is true (applies to all workspaces)
        sync:
          $ref: '#/components/schemas/FilterRuleSyncSummaryDto'
        workspaceSync:
          type: array
          items:
            $ref: '#/components/schemas/FilterRuleWorkspaceSyncDto'
          description: Per-workspace agent-manager sync status for this rule
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateFilterRuleDto:
      type: object
      required: [pattern, direction, filterType, isGlobal]
      properties:
        pattern:
          type: string
        regexFlags:
          type: string
        direction:
          type: string
          enum: [incoming, outgoing, bidirectional]
        filterType:
          type: string
          enum: [none, filter, drop]
        replaceContent:
          type: string
        priority:
          type: integer
          minimum: 0
        enabled:
          type: boolean
        isGlobal:
          type: boolean
        workspaceIds:
          type: array
          items:
            type: string
            format: uuid
    UpdateFilterRuleDto:
      type: object
      properties:
        pattern:
          type: string
        regexFlags:
          type: string
        direction:
          type: string
          enum: [incoming, outgoing, bidirectional]
        filterType:
          type: string
          enum: [none, filter, drop]
        replaceContent:
          type: string
        priority:
          type: integer
          minimum: 0
        enabled:
          type: boolean
        isGlobal:
          type: boolean
        workspaceIds:
          type: array
          items:
            type: string
            format: uuid
    AtlassianSiteConnectionResponseDto:
      type: object
      description: Connection metadata only; API tokens are stored encrypted at rest and are never returned.
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
        baseUrl:
          type: string
        accountEmail:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateAtlassianSiteConnectionDto:
      type: object
      required: [baseUrl, accountEmail, apiToken]
      properties:
        label:
          type: string
        baseUrl:
          type: string
          format: uri
        accountEmail:
          type: string
          format: email
        apiToken:
          type: string
          description: Atlassian API token; persisted with AES-256-GCM encryption (never returned on GET list/get connection).
    UpdateAtlassianSiteConnectionDto:
      type: object
      properties:
        label:
          type: string
        baseUrl:
          type: string
          format: uri
        accountEmail:
          type: string
          format: email
        apiToken:
          type: string
          description: When omitted, token is unchanged. If set, stored with AES-256-GCM encryption.
    AtlassianConnectionTestResultDto:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
    ExternalImportProviderId:
      type: string
      enum: [atlassian]
    ExternalImportKind:
      type: string
      enum: [jira, confluence]
    ExternalImportConfigResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        provider:
          $ref: '#/components/schemas/ExternalImportProviderId'
        importKind:
          $ref: '#/components/schemas/ExternalImportKind'
        atlassianConnectionId:
          type: string
          format: uuid
        clientId:
          type: string
          format: uuid
        enabled:
          type: boolean
        jiraBoardId:
          type: integer
        jql:
          type: string
        importTargetTicketStatus:
          $ref: '#/components/schemas/TicketStatus'
        confluenceSpaceKey:
          type: string
        confluenceRootPageId:
          type: string
        cql:
          type: string
        agenstraParentTicketId:
          type: string
          format: uuid
        agenstraParentFolderId:
          type: string
          format: uuid
        lastRunAt:
          type: string
          format: date-time
        lastError:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateExternalImportConfigDto:
      type: object
      required: [provider, importKind, atlassianConnectionId, clientId]
      properties:
        provider:
          $ref: '#/components/schemas/ExternalImportProviderId'
        importKind:
          $ref: '#/components/schemas/ExternalImportKind'
        atlassianConnectionId:
          type: string
          format: uuid
        clientId:
          type: string
          format: uuid
        enabled:
          type: boolean
        jiraBoardId:
          type: integer
        jql:
          type: string
        importTargetTicketStatus:
          $ref: '#/components/schemas/TicketStatus'
        confluenceSpaceKey:
          type: string
        confluenceRootPageId:
          type: string
        cql:
          type: string
        agenstraParentTicketId:
          type: string
          format: uuid
        agenstraParentFolderId:
          type: string
          format: uuid
    UpdateExternalImportConfigDto:
      type: object
      properties:
        atlassianConnectionId:
          type: string
          format: uuid
        clientId:
          type: string
          format: uuid
        enabled:
          type: boolean
        jiraBoardId:
          type: integer
        jql:
          type: string
        importTargetTicketStatus:
          $ref: '#/components/schemas/TicketStatus'
        confluenceSpaceKey:
          type: string
        confluenceRootPageId:
          type: string
        cql:
          type: string
        agenstraParentTicketId:
          type: string
          format: uuid
        agenstraParentFolderId:
          type: string
          format: uuid
...